server: wl_client: introduce method for getting client security context
Introduces a function wl_client_get_security_context for obtaining the
security context of a connected client by using SO_PEERSEC on supported
systems. Returns NULL for unsupported systems (inline with behaviour of
e.g. Linux attempting to get a security context for a client that does
not exist, which is a valid state).

Signed-off-by: Rahul Sandhu <nvraxn@gmail.com>