protocol: Make DnD pointer cursor updates responsibility of compositors
Originally, having the drag source client in charge of cursor updates
was simple enough, flexible and coherent with the prior art.

Things accumulated over time that made it not such a great choice:

- The tablet protocol sporting its own on-screen pointer cursors
  made clients need to be painfully aware of the device that started
  drag-and-drop, in order to modify the correct cursor, provide the
  correct serials, etc.
- The shapes protocol greatly simplified the set of sensible cursors
- On top of that, not every compositor nor toolkit library knows
  to do well all of that.

Nowadays, it seems simpler to leave this bit of user feedback up
to the compositor, as it will/should be aware of the device driving
drag-and-drop, as well as the currently selected action.

In order to allow for a seamless transition (mostly in the client
side), document this as mandatory only if version 4 of the
wl_data_* interfaces is implemented by the compositor. The version
bump only identifies to clients the possibility to avoid doing
any cursor feedback themselves.

Signed-off-by: Carlos Garnacho <carlosg@gnome.org>