protocol: specify exact multiplanar layout for wl_shm
This change calculates multiplanar buffer plane strides so that,
if the first plane is tightly packed, the other planes are also
tightly packed. Matching Vulkan's constraints on multiplanar formats,
it requires the width/height/stride parameters are divisible as
neccesary to avoid ever needing to round subsampled pixels.

This is technically a breaking change, but very few clients and
and compositors implemented and used multiplanar shm formats.
For a given format, those that do either agree with the new
calculations or disagree with each other.

Signed-off-by: Manuel Stoeckl <code@mstoeckl.com>