protocol: remove note discouraging use of wl_surface.damage
If a client tracks damage in surface coordinates, there is no reason for
it to use the damage_buffer request.

The message of the commit that originally added this note says that the
compositor might implement wl_surface.damage inefficiently and that it
is complicated to implement correctly. Having the logic replicated in
every client that tracks damage in surface coordinates does not make it
more efficient or more correct. The conversion between surface and
buffer damage can be implemented as a matrix multiplication but it's not
easy to get right and it is more likely that compositors will have a
correct and fast implementation.

Reverts: 921d0548035673a1bf6aeb9396b9bc728133411e
Signed-off-by: Julian Orth <ju.orth@gmail.com>